Fix std::accumulate overflow

+++ mindspore/ops/kernel/ascend/aclnn/kernel_mod_impl/customize/argmin_ext_aclnn_kernel.cc
@@ -43,8 +43,8 @@ void ArgMinAscend::GetWorkSpaceInfo(const std::vector<KernelTensor *> &inputs,
   if (dim_is_none_) {
     input_kernel_tensor_ = inputs[kIndex0]->CloneKernelTensor();
 
-    int input_flatten_size =
-      std::accumulate(input_real_shape_.begin(), input_real_shape_.end(), 1, std::multiplies<int64_t>());
+    int input_flatten_size = std::accumulate(input_real_shape_.begin(), input_real_shape_.end(),
+                                             static_cast<int64_t>(1), std::multiplies<int64_t>());
     auto input_flatten_shape = ShapeVector{input_flatten_size};
     input_kernel_tensor_->SetShapeVector(input_flatten_shape);
