+++ mindspore/ops/kernel/gpu/cuda/other/hamming_window_gpu_kernel.cc
@@ -54,7 +54,8 @@ int HammingWindowGpuKernelMod::Resize(const std::vector<KernelTensor *> &inputs,
   MS_ERROR_IF_NULL_W_RET_VAL(output, KRET_RESIZE_FAILED);
   std::vector<int64_t> output_shape =
     std::vector<int64_t>(output->GetDeviceShapeVector().begin(), output->GetDeviceShapeVector().end());
-  output_elements_ = std::accumulate(output_shape.begin(), output_shape.end(), 1, std::multiplies<int64_t>());
+  output_elements_ =
+    std::accumulate(output_shape.begin(), output_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
   if (output_elements_ == 0) {
     is_null_input_ = true;
   }
